Add Jest for testing
