+++ components/Plot.jsx
