+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    moduleFileExtensions: ["js", "jsx"],
+    testMatch: ["**/tests/**/*.test.js?(x)"],
+    testPathIgnorePatterns: ["<rootDir>/build/", "<rootDir>/node_modules/"],
+}
