+++ package.json
@@ -11,12 +11,13 @@
     "keywords": [],
     "license": "AGPL 3",
     "scripts": {
-        "build": "next build app",
-        "start": "npm run build && env NODE_ENV=production node server.js",
-        "dev": "nodemon --ignore app/ server.js",
+        "build": "next build",
+        "dev": "nodemon --watch server.js server.js",
+        "prod": "npm run build && env NODE_ENV=production node server.js",
         "style:check": "eslint -c .eslintrc --ext .js,.jsx --ignore-path .gitignore .",
         "style:format": "npm run style:check -- --fix",
-        "style:rules": "eslint --print-config .eslintrc"
+        "style:rules": "eslint --print-config .eslintrc",
+        "test": "jest -c jest.config.js"
     },
     "dependencies": {
         "@material-ui/core": "^3.9.3",
@@ -42,6 +43,7 @@
         "@babel/plugin-proposal-pipeline-operator": "^7.3.2",
         "@babel/preset-flow": "^7.0.0",
         "babel-eslint": "^10.0.1",
+        "babel-jest": "^24.8.0",
         "babel-plugin-flow-react-proptypes": "^25.1.0",
         "babel-plugin-module-resolver": "^3.2.0",
         "babel-plugin-wrap-in-js": "^1.1.1",
@@ -50,8 +52,10 @@
         "eslint-plugin-flowtype": "^3.7.0",
         "eslint-plugin-fp": "^2.3.0",
         "eslint-plugin-import": "^2.17.2",
+        "eslint-plugin-jest": "^22.5.1",
         "eslint-plugin-jsx-a11y": "^6.2.1",
         "eslint-plugin-react": "^7.13.0",
+        "jest": "^24.8.0",
         "nodemon": "^1.19.0"
     }
 }
