+++ styles/theme.js
@@ -1,0 +1,9 @@
+import { createMuiTheme } from "@material-ui/core/styles"
+
+const theme = createMuiTheme({
+    typography: {
+        useNextVariants: true,
+    },
+})
+
+export default theme
