+++ tests/lib/getPageContext.test.js
@@ -1,0 +1,8 @@
+/* eslint-env jest */
+
+import getPageContext from "./../../lib/getPageContext"
+
+test("it works", () => {
+    const pageContext = getPageContext()
+    return expect(pageContext).toBe(3)
+})
